啊啊啊

+++ src/main/java/xmu/crms/serviceImpl/TopicServiceImpl.java
@@ -24,41 +24,83 @@ public class TopicServiceImpl implements TopicService
 	private TopicMapper topicMapper;
 	@Override
 	public Topic getTopicByTopicId(BigInteger topicId) throws TopicNotFoundException, IllegalArgumentException {
-		return topicMapper.getTopicByTopicId(topicId);
+		 if(topicId.intValue()<=0)
+         {
+             throw new IllegalArgumentException();
+         }
+		Topic topic = topicMapper.getTopicByTopicId(topicId);
+		if(topic==null)
+		{
+			throw new TopicNotFoundException();
+		}
+		return topic;
 	}
 
 	@Override
-	public void updateTopicByTopicId(BigInteger topicId, Topic topic)
-			throws TopicNotFoundException, IllegalArgumentException {
-		topicMapper.updateTopicByTopicId(topicId, topic);
+	public void updateTopicByTopicId(BigInteger topicId, Topic topic)throws TopicNotFoundException, IllegalArgumentException {
+		 if(topicId.intValue()<=0)
+         {
+             throw new IllegalArgumentException();
+         }
+		Integer num=topicMapper.updateTopicByTopicId(topicId, topic);
+		if(num==null)
+		{
+			throw new TopicNotFoundException();
+		}
 		
 	}
 
 	@Override
-	public void deleteTopicByTopicId(BigInteger topicId) throws IllegalArgumentException {
-		topicMapper.deleteTopicByTopicId(topicId);
+	public void deleteTopicByTopicId(BigInteger topicId) throws IllegalArgumentException, TopicNotFoundException {
+		 if(topicId.intValue()<=0)
+         {
+             throw new IllegalArgumentException();
+         }
+		Integer num=topicMapper.deleteTopicByTopicId(topicId);
+		if(num==null)
+		{
+			throw new TopicNotFoundException();
+		}
 	}
 
 	@Override
 	public List<Topic> listTopicBySeminarId(BigInteger seminarId) throws IllegalArgumentException {
-		
+		 if(seminarId.intValue()<=0)
+         {
+             throw new IllegalArgumentException();
+         }
 		return topicMapper.listTopicBySeminarId(seminarId);
 	}
 
 	@Override
 	public BigInteger insertTopicBySeminarId(BigInteger seminarId, Topic topic) throws IllegalArgumentException {
-		topicMapper.insertTopicBySeminarId(seminarId, topic);
-		return topic.getId();
+		 if(seminarId.intValue()<=0)
+         {
+             throw new IllegalArgumentException();
+         }
+		Integer num=topicMapper.insertTopicBySeminarId(seminarId, topic);
+		if(num<=0)
+		{
+			return new BigInteger("-1");
+		}
+		else return topic.getId();
 	}
 
 	@Override
 	public void deleteSeminarGroupTopicById(BigInteger groupId, BigInteger topicId) throws IllegalArgumentException {
+		if(groupId.intValue()<=0||topicId.intValue()<=0)
+        {
+            throw new IllegalArgumentException();
+        }
 		topicMapper.deleteSeminarGroupTopicById(groupId, topicId);
-		
 	}
 
 	@Override
 	public void deleteSeminarGroupTopicByTopicId(BigInteger topicId) throws IllegalArgumentException {
+		if(topicId.intValue()<=0)
+        {
+            throw new IllegalArgumentException();
+        }
 		topicMapper.deleteSeminarGroupTopicByTopicId(topicId);
 		
 	}
@@ -66,12 +108,19 @@ public class TopicServiceImpl implements TopicService
 	@Override
 	public SeminarGroupTopic getSeminarGroupTopicById(BigInteger topicId, BigInteger groupId)
 			throws IllegalArgumentException {
-		
+		if(groupId.intValue()<=0||topicId.intValue()<=0)
+        {
+            throw new IllegalArgumentException();
+        }
 		return topicMapper.getSeminarGroupTopicById(topicId, groupId);
 	}
 
 	@Override
 	public void deleteTopicBySeminarId(BigInteger seminarId) throws IllegalArgumentException {
+		if(seminarId.intValue()<=0)
+        {
+            throw new IllegalArgumentException();
+        }
 		List<Topic> topics=topicMapper.listTopicBySeminarId(seminarId);
 		for(int i=0;i<topics.size();i++)
 		{
@@ -83,8 +132,12 @@ public class TopicServiceImpl implements TopicService
 
 	@Override
 	public List<SeminarGroupTopic> listSeminarGroupTopicByGroupId(BigInteger groupId) throws IllegalArgumentException {
-		
-		return topicMapper.getSeminarGroupTopiclistByGroupId(groupId);
+		if(groupId.intValue()<=0)
+        {
+            throw new IllegalArgumentException();
+        }
+		List<SeminarGroupTopic> seminarGroupTopiclist=topicMapper.getSeminarGroupTopiclistByGroupId(groupId);
+		return seminarGroupTopiclist ;
 	}
 
 }
